requirements.txt
pytest
ruff
black
mypy
python-dotenv
PyQt6
google-generativeai
